删除多余示例
