--- src-client/pages/admin/components/Bar.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function Bar() {
-  return <div>And I am Bar!</div>
-}
