--- src-client/pages/admin/components/Foo.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function Foo() {
-  return <div>I am Foo!</div>
-}
