--- src-client/pages/admin/components/index.js
@@ -1,4 +1,0 @@
-export App from './App'
-export Home from './Home'
-export Foo from './Foo'
-export Bar from './Bar'
